TileSimAlgs: Fix multiple hits for the same channel for Run 1 (ATLASSIM-6608)
TileSimAlgs: Fix multiple hits for the same channel for Run 1 (ATLASSIM-6608)

TileHitVecToCntTool tool merges multiple hits for the same channel
in the output TileHitContainer. The multiple hits are expected to come
from Fast Calorimeter simulation, but in the case of merged E1,MBTS
they also look as multiple hits for the same channel in Full G4 Simulation.
So, there is special case during multiple hits merging to treat them correctly
in order to keep the TileHitContainer unchanged in this case.

But there were no merged E1,MBTS in Run 1, so protection has been added
to this special case to work only in Run 2+ geometries and for modules
in which merged E1,MBTS actually exist accordingly to the Tile cabling.