tauRec: early stopping of tau reconstruction if tau fails AOD thinning pt cut
tauRec: early stopping of tau reconstruction if tau fails AOD thinning pt cut

Hello,

This MR is implementing an "early" stopping in the tau reconstruction: when a tau has an MVA-calibrated pt below the AOD thinning threshold, we now skip the last few tools scheduled in the tau reconstruction (mostly identification tools), to save CPU.
I've checked the data18 SPOT test, TauRecRunnerAlg is now 12% faster, the gain is modest as the skipping requires MVA pt calibration, which happens late in the tau reconstruction. The AOD content is unchanged.

In principle, this change should not affect anything, as we are not really supposed to use or even monitor taus (e.g. DQ monitoring in RAWtoALL) below the AOD pt thinning threshold, as these taus are transient and not written to AOD and therefore not used for physics.
If the monitoring changes (e.g. tau ID scores becoming zero as no longer computed for soft taus), the monitoring should be instructed to skip taus failing the thinning AOD cut.

Cheers,
Bertrand