TrigConfHLTUtils: performance improvement to string2hash
TrigConfHLTUtils: performance improvement to string2hash

Migrate `string2hash` from TBB containers to the concurrent maps
in CxxUtils, which provide much better performance for our use-case.

In addition to the hash->string mapping, also store the reverse
string->hash so we do not need to re-calculate the hash on each lookup.