Removing muon quality variable from PHYSLITE
Following the discussion in https://its.cern.ch/jira/browse/ATLASG-2544
, it seems that muon quality as written directly is not something that
should be used for analysis. If that's the case, then I (personally)
find it more clear to remove the variable to make it obvious that it
cannot be used.

Just offering an option here to make clear that this is a
(straightforward) technical option.

This shouldn't get merged without some comment from @mvanadia ,
@hrussell , @gwatts , @egramsta (or some combination of them), but not
marking as draft so that the CI goes ahead and runs.

This is a duplicate of !63787 but pointing at the main branch.