TrigValTools: update data23_cos input
TrigValTools: update data23_cos input

Hello,

This MR is updating the data23_cos RAW data files used in trigger ART tests. Fortunately, this time, we have enough stats (we process up to 4k events) from a same LB, so the issue reported in ATR-26472 wouldn't show up if we were to use a grid dataset instead of files on cvmfs (for simplicity, we stick to cvmfs inputs).

Cheers,
Bertrand