TrigInDetValidation: increase TrigInDetRdictStep timeout
TrigInDetValidation: increase TrigInDetRdictStep timeout

Hello,

This MR is increasing the timeout of the TrigInDetRdictStep in trigger ART tests from 10 to 15 mins, to avoid occasional timeout errors (ATR-27688).
For simplicity, the timeout is increased for all ID tests. One option could be to introduce an extra agument, to be able to specify the timeout only for a given test, but that seems a bit overkill.

Cheers,
Bertrand