GCC Clang Pragma Warning Fix, 23.0 branch (2023.07.10.)
Only allow Cling to see the warning suppression pragmas.

As it turns out, GCC is very touchy about seeing pragmas that are
not meant for it. So we need to only expose these pragmas to the
ROOT dictionary generation. Even Clang as the main C++ compiler
does not need to see them.