TrigInDetValidation: update ParticleGun RDO inputs
TrigInDetValidation: update ParticleGun inputs

Hello,

This MR is updating the ParticleGun RDO input files used in TrigInDetValidation ART tests. We now use an MC23c setup!
As discussed in ATLMCPROD-10688, we now have two more samples for which we'll need new tests:
* 901970/mc.PG_mu_Pt4to100.py without pileup, grid dataset = valid1.901970.PG_mu_Pt4to100.recon.RDO.e8544_e8528_s4159_s4114_r14813_tid34085598_00 with 2 files of 10k each, cvmfs input = "Single_mu"
* 901970/mc.PG_mu_Pt4to100.py with pileup, grid dataset = valid1.901970.PG_mu_Pt4to100.recon.RDO.e8544_e8528_s4159_s4114_r14800_tid34086072_00 with 10 files of 2k each, cvmfs input = "Single_mu_pu"

I have not created the new tests, as I was not 100% sure what cuts should be applied in TrigInDetAnalysisUser, and will let experts take care of that.

Cheers,
Bertrand