Daily merge of 23.0 into main
TrigJetHypo ATR-27619 - reduce the number of AlgTools.

TrigJetHypo ATR-27619 - reduce the number of AlgTools.

Replace the on to one matching of Condition filters - Conditions with an
index array, and a Filter array. The index array is one-to-one with the Conditions,
while the size of the filter array is at most the length of the filter index array.

In the majority of cases, the condition filters were pass-through filters, so this
approch reduces the need to configure the such filters, at the proce of more complex code.

Further, the filter used before the processing the jet hypo, has been replaced by a simple copy
when a PassThrough filter was requested.