Update trigger cosmic tests to not run L1 Simulation on data
Update trigger cosmic tests to not run L1 Simulation on data

As discussed in ATR-27976, the L1 muon simulation on cosmics data is very inefficient leading to very low statistics passing all stages of the cosmic HLT chains. To get around this, we're turning off the L1 simualtion  in the cosmic data trigger tests and instead setting forceEnableAllChains to true so that all the HLT chains will run in every event (to get aorund the fact that the L1 decisions in data may not match with the current L1 menu in the release). Removing as well the rewriteLVL1 option for the trigID cosmic data, as this isn't needed if we aren't running the L1 simulation.