Fix JetSuper Roi constituent management
Fix JetSuper Roi constituent management

The JetSuper Roi was being told to *not* manage it's constituents, but then
mistakenly, unique_ptr was being used to created those constituents, but of course,
as they were being stored as standard pointers, they were being released first,
so the unique pointer was no longer managine them, and they were no longer being
managed by unique_ptr and were subsequently leaking when the superRoi was being
destroyed.

In principle the unique_ptr could have then been reset() with the now, standard
ptr, but then the now standard constitent pointer i the Roi would not have been
safe to use once the unique_ptr had gone out of scope  - it would have still be
there, but could the memory be treclaimed at some later stage ? So would it be safe
to use ? I suspect not, and almost certainly calling delete on it would cause a
malloc error as the address would have already been deleted when the unique_ptr
went out of scope.

As such unique_ptr are in appropriate here, and so we resort back to good old "new"
and let the Roi manage it's consituents itself.

All Roi copies are deep copies of the constituents now, so this is guaranteed safe.