CaloClusterMatching: Fix use of WriteDecorHandle.
CaloClusterMatching: Fix use of WriteDecorHandle.

When a WriteDecorHandle is destroyed, the decoration gets locked.
Therefore, the Handle needs to live until all decorations are written.

(There was a bug that meant that the container sometimes did not appear
to be locked, which is why this was working before.   Need to clean this up
in order to actually fix that bug.  See ATEAM-909.)