AtlasBuildScripts: fix bug in logic to decide on externals rebuild
Rewrite the code to decide if an externals rebuild is needed. Instead of
using a combination of the externals version number and a hash of all
(c)make arguments, simply concatenate all relevant information into one
big "stamp" file.

This fixes also the problem of re-using stale builds (ATLINFR-5026).

Relates to ATLINFR-3018.