TrigEDMConfig: fix bug in checking view collections
The `getRun3TrigObjProducedInView` method only did a partial match of the
collection key when checking if a collection is a view collection. E.g.
`HLT_TopoCaloClustersLC` would match to `HLT_TopoCaloClustersLCFS`
because it comes first in the list.

For the v1PhysP1 menu this results in the following change to the
`CollectionsToSerialize` property:

```
< 'xAOD::CaloClusterTrigAuxContainer#HLT_TopoCaloClustersLCAux.nCells.CENTER_MAG;0'
---
>  'xAOD::CaloClusterTrigAuxContainer#HLT_TopoCaloClustersLCAux.nCells.CENTER_MAG.viewIndex;0'
```