AthenaConfiguration: fix orphaned tools in legacy support
When renaming a configurable one needs to ensure to also set the
`_jobOptName` slot (where available). Otherwise this leads to orphaned
configurables instances without a parent. These do not create a problem
but appear needlessly in the final pickle file.

Also ensure that the global Configurable DB is adjusted correctly after
renaming.