Moving the CA wrappers for the muon trigger inside-out sequence up
Moving the CA wrappers for the muon trigger inside-out sequence up

Changing the muon trigger inside-out algorithm sequence to return a CA and then adding the algorithmCAToGlobalWrapper around the algorithm sequence when it is called as part of teh menu sequence. Ongoing CA migration of precision muon trigger steps (ATR-25086).