TrigDecision+HLTResult: decode ROB truncation status
Decode the per-ROB truncation status and store it in the `TrigDecision`.
Since we only support one truncation status, we use the convention of
reflecting the main HLT Result ROB if available, otherwise the combined
status of all other ROBs (but there should only be one for a given
non-Physics stream).

In any case, for severe truncations (debug stream) the status from the
event header has priority.

Companion to !64583 and relates to ATR-27986.