AthenaCommon: Update for py 3.11.
AthenaCommon: Update for py 3.11.

Reduce the verbosity of unittest to the default.
The python unit tests were run with the verbosity option set to 2,
which meant that there was printout for each individual test.
However, in py 3.11, the format that unittest uses to print test names changes.
Deal with this by just reverting the verbosity setting to the default.
Update the reference files accordingly.
does.