provide a default bevavior if GridWorker is run on unknown datasets
In the existing code it tries to read the sample metadata based on the
dataset.  This works fine, but breaks for central n-tuple production,
as that wants to run on new datasets that weren't known at submission.
I now added a default behavior which is hopefully sufficient for the
purpose.