L1CaloFEXSim: cleanup unused headers
L1CaloFEXSim: cleanup unused headers

Hello,

This MR is mostly dropping headers that are not used, in the eFEX domain. Some sides changes are made to prepare for algorithm re-entrancy (e.g. turn methods to const).
More to come, I submit this batch not to make the MR too big.

Cheers,
Bertrand