allow enabling EventLoop grid error reporting outside of GridDriver/PrunDriver
We are trying to make it possible to run user n-tuple production as
part of central production.  For that it is very helpful if the jobs
do the proper error reporting.  For GridDriver and PrunDriver that is
done by default, and that is the recommended way to do it.  However,
some users have their own wrapper scripts and use DirectDriver on the
grid.  Those users can now enable the grid reporting functionality in
their jobs as well.