CITest: add MC trigger test with HI menu
CITest: add MC trigger test with HI menu

Hello,

This MR is adding a new trigger test to the Athena CI, to test the Heavy Ion trigger menu and improve stability as HI data taking approaches.
The test was also added to the minimal set of tests executed when running `runTrigART.py -m`.
The test currently uses ttbar RDO as input, this may change if we manage to produce a Hijing RDO (ATLHI-483).

Cheers,
Bertrand