Make FrozenShowers compatible with HepMC3
Make FrozenShowers compatible with HepMC3

1. Adding missing destructor (relevant also for HepMC2).
A FastSimModel was initialised in FastSimulationBase.cxx but
was never deleted. As a result, a HepMC file created by a
fast simulation model in LArFastShower.cxx closed improperly.
A destructor is created in FastSimulationBase
instead of the default one to fix the issue.

2. Making FrozenShowers compatible both with HepMC2 and HepMC3.
HepMC3 methods are added. Also, a dummy input particle is added
so that HepMC record can be saved saved.

3. Making units explicitly specified to avoid confusion.
Also, the hardcoded GeVToMeV units conversion in
HepMCReadFromFile.cxx is changed to a flexible conversion method.

4. The HepMC record filtering macro is updated with python bindings.
However, to allow reading old HepMC files, an old macro is renamed to
LArG4FSStartPointFilterLegacy.py and kept for the moment.

5. Adding --runNumber option in FrozenShower library
production jobOptions to avoid occasional detector mis-alignment.