Fixing hadron origin classifier/decorator configuration
Fixing hadron origin classifier/decorator configuration

The configuration of tools in the script HFHadronsCommonConfig was incompletely migrated to the CA and was completely broken. This was never noticed because the list of DSIDs for which the tool is run hasn't yet been updated, so the config was never used. This MR fixes the first part of the issue - the configuration is now complete. This only has an effect for the rarely used workflow AOD->TRUTH1, which is never used in production, but which was broken by this issue. For the more commonly used EVNT->TRUTH1 this fix will have no effect until the DSIDs are updated, to be followed up.