add option to check for missing columns in compareFlatTrees
During my previous changes I had to realize that our tests don't flag
columns disappearing, which is not ideal.  This now adds the option to
check for that as well.  I added that as an option so as not to change
existing behavior.

I enabled that for the CP algorithm checks, but had to disable some
columns that the sequence configuration didn't have at all or had
different results.  I'm not sure if it is worth fixing this (given
that the sequence configuration is on its way out), but if so this
should be part of a future change (given that it currently also
doesn't get checked).