DQUtils: Don't load oracle if unneeded.
DQUtils: Don't load oracle if unneeded.

Avoid loading the oracle module if it's not actually going to be used.

Fixes a test failure in DCSCalculator2 seen on el9.
[The failure that was seen was due to an interface change in the sqlalchemy
library used by the oracle module.  I first tried fixing that up.
But apparently the old interface would open the database collection lazily
while the new interface would open it immediately.  And opening the
database connection to oracle didn't work.  But the oracle connection wasn't
actually used in this test, so rather than try to figure out what's going
wrong with the oracle connection (which may just not work anymore due
to DB changes), i just changed things so as not to load the oracle
module if it's not needed. If it ever is actually needed, the sqlalchemy
changes will still need to be dealt with.]