TriggerJobOpts: add module for postExec modifiers
Add a new `PostExec` module that should host any code that can be used
with `runHLT.py` via `athenaHLT -C` or `athena --CA --postExec`.

The only use-case so far is the `forceConditions` postExec that should
be used as follows:
```
athenaHLT.py -C 'from TriggerJobOpts import PostExec; PostExec.forceConditions(440499,470)' ...
```
Note that we no longer provide access to run and LB number so those have
to be specified explicitly based on the input file.

The legacy `forceConditions` Modifier re-uses the new code.