TrigAnalysisTest: increase test timeout
TrigAnalysisTest: increase test timeout

Hello,

This MR is increasing the timeout from 3600 (default) to 4200 seconds, for a couple of trigger ART tests that are timing out (ATR-28113).
For build tests, given that most of the time is spent in the initialisation rather than in the event processing, it seems more appropriate to increase the timeout than decrease the number of events (20 by default).

Cheers,
Bertrand