CP::ShowerDepthTool Updates, main branch (2023.08.21.)
Cleaned the interface and implementation of CP::ShowerDepthTool.

Primitive types are now passed to the tool by value instead of
const reference.

Made the tool use std::unique_ptr for managing the memory of
its histograms.

Re-wrote the protection against out-of-bounds interpolation in
the code to not hardcode any properties of the underlying TH2
histogram.