xAODMissingET: Protect agaisnt out-of-bounds shift.
In MissingETAssociationHelper::objSelected, check that objIdx is not
outside of the possible bounds for a bit shift.
This can happen in the objSelected overload taking an IParticle* is
called, and the particle is not found.  In that case, objIdx will be -1.

Cleans up an error seen from the undefined behavior sanitizer.