ComponentAccumulator: avoid unmerged CA errors in case of unrelated problems
Avoid checking for unmerged CAs in case of unrelated problems. This
typically happens when an exception is thrown and the job is aborted.
The use of `sys.exc_info` no longer works for this purpose in Python3.
Use an exit-handler instead.

This open a small loophole in top-level scripts, e.g.:
```
  cfg = ComponentAccumulator()
  cfg.addEventAlgo(CompFactory.HelloAlg())
```
would no longer create an "unmerged" error. On the other hand we are no
longer forced to add the `cfg.wasMerged()` in all unit tests.

Also slightly improve the error message itself.