CaloRec: Fix test failure on arm.
CaloRec: Fix test failure on arm.

CaloCellMakerTest was comparing a dump of the cells with a reference file.
This was observed to be failing on aarch64 because FP results are slightly
different.

Now, incorporate the functionality of comparing with a reference into the
dumper algorithm; in that case, we can use inexact comparison for the FP values.