Porting MUON1 to component accumulator
Draft: porting MUON1 to component accumulator

This MR provides a port of the MUON1 derivation to the component accumulator. It is intended that this will be used as a template for the other MCP derivations. As well as the main config fragment, the MR also provides several new config fragments for certain common tools that haven't been previous provided.

**Not yet working** as several of the config fragments still have to be written