CP Algorithms: overhaul of the handling of overlap removal selections
The main thing this does is to introduce an AsgUnionPreselectionAlg
that allows to combine systematics dependent selection flags into a
uniform preselection per systematic copy.  The logic here is that in
preselections I (typically) just want to filter out objects for
computational efficiency, but if the object fails for one systematic
on the shallow copy I anyways have to process it.  Having such a joint
selection decoration prevents the systematics of the selection to
propagate into outputs that would otherwise be unaffected by it.

This then in turn allows me to add such a decoration (on request) as
part of the overlap removal config, as that is the only config block
that has systematics dependent selection decorations.

I also added a bunch of options to the OverlapRemovalConfig to allow
changing the selection name to whatever the user wants/needs.