L1CaloFEXAlgos: Fix usage of WriteDecorHandle.
L1CaloFEXAlgos: Fix usage of WriteDecorHandle.

When a WriteDecorHandle is deleted, the decoration gets locked.
So need to move the WDH declarations out of the element loop, so they're 
only created/destroyed once.

This was previously working due to a bug that resulted in decorations sometimes
appearing not to be locked.  See ATEAM-909.