IdDictDetDescrCnv: implemented a mechanism for reading ID dictionaries from SQLite as BLOBs
If GeoModel is initialized from SQLite, then IdDictDetDescrCnv first attempts to read
ID dictionaries from SQLite tables in BLOB format. In case of success, the dictionaries
will be written out as XML files into the run directory and passed to the dictionary
parser for further processing.

This can be considered an interim solution aimed to avoid making any changes in the
ID dictionary parser, and in this way to support several options for ID dictionary
initialization:
1. Retrieve ID dictionary file names from Geometry DB tables
2. Retrieve ID dictionary file names from SQLite database tables
3. Retrieve ID dictionaries as BLOBs from SQLite database tables

Later we can consider dropping the step of writing BLOBs on disk as XML files, and
instead passing them as strings to the dictionary parser.