disable the block configuration tests for 21.2
In 21.2 we never made the block configuration fully agree with the
sequence configuration and we stopped developing it, so I disabled the
corresponding tests.  Leaving the tests commented out since this
branch is on its way out and if somebody tries to investigate what's
going on I want to leave them with a trace of it.