Ppr/PPM/CPMMonitoring: Fix wrong variable type for event numbers and bug related to mismatch-event-counter
Ppr/PPM/CPMMonitoring: Fix wrong variable type for event numbers.

The event number variable used in the Ppr-,PPM, and CPM-Monitoring was inialized as integer, whereas it should be long long.
This lead to negative event numbers (2-complement) or none at all.

The bug was fixed by changing the type.