VP1AODSystems: slots -> Q_SLOTS
Rename the Qt slots 'keyword' to Q_SLOTS.

Having this in the source confuses cppcheck, which reports syntax errors.
We can define slots as a macro expanding to nothing in the cppcheck
configuration, but if we do that globally, there are other packages that
break because they use slots as an identifier.  Using Q_SLOTS instead
prevents this name collision.