MuonMDT_CnvTools: Improve handling of empty ROD.
MuonMDT_CnvTools: Improve handling of empty ROD.

If we had a ROD with no payload; that is, a BOB word immediately followed
by an EOB word, then MdtROD_Decoder would return a failure status.
However, this has actaully occurred in data, and was causing the
test_bulkProcessing_pilotBeam_local.sh ART test to fail.

Change so that in that case we return success (but print a warning).

Also improve diagnostics.  We should log an ERROR whenever we fail, and
since a number of those messages are identical, include the line number
in the message as well.

Also include a reference to the note documenting the MDT data format.