Only run Pixel, SCT and TRT conversion to xAOD for TIDE preselected events.
The conversion of Pixel, SCT and TRT raw data to xAOD is very time
consuming. Since the content is only written to the IDTIDE stream
these algorithms only have to run for IDTIDE selected events.

This is realised by extra sequences. There is now one sequence
operated sequentially until an algorithm (i.e. the IDTIDEKernelPresel
or IDTIDEKernel) set the filter status to reject the event. Then
there is a sequence (IDTIDEPreselAlgSequence) for algorithms like the
conversion algorithms which are executed after IDTIDEKernelPresel in
parallel. Finally there is a sequence (IDTIDEPostProcSequence) for the
augmentation and thinning kernels which are run after the final selection
of the IDTIDEKernel (also in parallel):

 - IDTIDEPreselSequence (sequential,
                         stop depending on algorithm filter status)
   - DerivationKernel/IDTIDEKernelPresel
   - IDTIDEPreselAlgSequence (parallel)
      - TRT_PrepDataToxAOD/xAOD_TRT_PrepDataToxAOD
      - SCT_PrepDataToxAOD/xAOD_SCT_PrepDataToxAOD
      - PixelChargeToTConversion/InDetPixelChargeToTConversion
      - PixelPrepDataToxAOD/xAOD_Pixel_PrepDataToxAOD
   - DerivationKernel/IDTIDEKernel
   - IDTIDEPostProcSequence (parallel)
      - DerivationKernel/DFTSOSKernel
      - DerivationKernel/IDTIDEThinningKernel