CpmSimMonitorAlgorithm: Fix simulation errors in the CP system due to disabled hadronic inputs during heavy ion running
CpmSimMonitorAlgorithm: Fix simulation errors in the CP system due to disabled hadronic inputs during heavy ion running

Due to a special configuration of the system the legacy hadronic inputs to the CP system need to be disabled during heavy ion configuration.
So far this was not reflected in the monitoring simulation of the CP system, which lead to transmission simulation errors appearing in the monitoring.

This commit fixes these errors by adding a flag to the CpmSimMonitoringAlgorithm that allows to mask the hadronic inputs in the
simulation by setting them to zero.