emergency fix for the errors introduced by DirectInputModule
Well, I guess that the lesson here should be always to write unit
tests when if statements are involved, not just rely on the
integration tests.  On th plus side the new infrastructure made it a
lot easier to test this, and I may have fixed a longer-standing bug
since the previous implementation had never been tested.