PanTau: replace legacy JobProperty flags with config flags
PanTau: replace legacy JobProperty flags with config flags

Hello,

This MR is replacing the JobProperty flags with config flags in the PanTau package.
This was blocking !61874 .
I ran diff-root on 1k events with standalone tau reco from ESD, the reconstruction output is hopefully unchanged.

It's a bit overkill to define so many flags, given that we never change their values, but a larger refactoring is beyond the scope of this MR.

Cheers,
Bertrand