Fix view name for muon l2io container in EDM config
Fix view name for muon l2io container in EDM config

As part of teh CA migration, one of the view names was forgotten to be updated for the L2 inside-out muon containers in the EDM config, causing the containers to be missing which is a problem for the monitoring that expects the containers to exists. Updating the view name to the correct version. Fixes ATR-28283.