Migrate latemu trigger seqeunce to MenuSquenceCA
Migrate latemu trigger seqeunce to MenuSquenceCA

Migrating the late muon sequence to be entirely CA based, and adding the MenuSequenceCAToGlobalWrapper to use the same sequence in legacy configuration. Note these sequences are not currently running in any trigger chains (migration for completeness so that they can eventually be used). Part of the ongoign work on ATR-25086.