add options to disable recalibrating objects from PHYSLITE
According to @gwatts it would be "amazing not to have to run
calibrations on PHYSLITE".  So here we go: This adds options to jets,
e-gamma, and muons to skip the calibration where possible.  This means
no calibration is run for jets, and for e-gamma and muons the nominal
calibration is skipped.  Either way, this is probably most noticable
when running with systematics turned off, but even then the savings
may be smaller than what MET alone costs us to run.

I did enable some comparison tests for PHYSLITE jobs that were
previously commented out (the data one failed, so I left it commented
out), but didn't enable the new feature until it is fully validated
(read: those tests did fail).  I'm leaving it to the PHYSLITE experts
to do the validation and then turning this on by default in a separate
MR.