ConfigText: Allow algorithms to be configured using a YAML configuration file
ConfigText: Allow algorithms to be configured using a YAML configuration file

ConfigTex.py contains the implementation to read, print, and process a YAML configuration
file. The contents of the configuration files will be used to configure the algorithms
used to define objects, OLR, define input and ouput.

An example configuration file has been added to the data directory to demonstrate a simple
analysis configuration as well as provide information on the available options and the
expected type.

The implementation requires the pyyaml package that is currently not included in
AnalysisBase releases and must be installed manually. The package has been requrested
in ATLINFR-4882.