TextConfiguration: Fix bug that would cause ftag blocks to be mis-configured...
TextConfiguration: Fix bug that would cause ftag blocks to be mis-configured when 'postfix' was not set.

When creating and configuring an instance of the FTag config block, if 'postfix'
was left as the empty string (i.e. ''), the block would not be configured properly.
To prevent this, the config sequence returned from 'makeConfig' is configured before
being added to the global config sequence that accumulates configured blocks.

A small function was also added that will print out all config blocks in a config
sequence and their configuration to help debugging similar issues in the future.