DerivationFrameworkPhysicsValidation: add TauTracks to AllVariables in DAOD_PHYSVAL
DerivationFrameworkPhysicsValidation: add TauTracks to AllVariables

Hello,

It seems we're not monitoring some of the main tau track parameters in PhysVal (track impact parameters + significances) as they are not part of the TauJets smart slimming list.
This results in empty plots in the PhysVal web display.
I'm adding TauTracks to `PHYSVALSlimmingHelper.AllVariables`. We're now getting the desired variables (plus a few we're not necessarily interested in).

Cheers,
Bertrand