Add new AthenaPool converter able to write old persistent version
Add new AthenaPool converter able to write old persistent version

This converter is able to write two peristent representations depending on
the value returned by a user-defined method writeOldPers(). Usually that
would be the current version and one of the old versions.
Given the Event data format used be default in Athena, the choice of
which version to write should be fixed for a given job.