Run Woodcock Tracking in the EMEC rather than EMECPara G4Region by default (ATLASSIM-6426)
Run Woodcock Tracking in the EMEC rather than EMECPara G4Region by default (ATLASSIM-6426)

Currently we use Woodcock tracking in the "EMECPara" `G4Region` which consists of a subset
of the volumes in the EMEC envelope. There is also an "EMEC" `G4Region` which contains the
top-level EMEC volumes. In order to avoid creating the "EMECPara" `G4Region` as this causes
some potential issues (ATLASSIM-6426), we would like to run Woodcock tracking in the "EMEC"
`G4Region` instead. This change was validated in ATLPHYSVAL-937.