CP Algorithm Block Config: Update postfix used for ftag, electron WP, etc. to selectionName.
CP Algorithms: Update postfix used for ftag, electron WP, etc. to selectionName.

The name postfix is used for many of the CP config blocks as second arguement to the
initialization but the postfix often corresponds to the name given to the selection
name in other CP blocks such as MET and overlap removal. Inorder to make the name
more meaningful, it has been changed to selectionName.